requirements.txt
fastapi
uvicorn
xgboost
scikit-learn
pandas
datasets
jinja2
python-multipart
pytest
httpx
twikit
python-dotenv